workspace/requirements.txt
chainer==6.0.0
Pillow==5.4.1
pyserial==3.4
jupyterlab==0.35.6
